lect_03/lect_03/requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
sqlparse==0.5.0
markdown2==2.4.0

